refactor: :recycle: Move files related to process execution to exec/
