+++ src/exec/util.c
@@ -21,4 +21,3 @@ void	receive_input_from_pipe(t_fd pipefd[PIPE_SIZE])
 	close(pipefd[PIPE_WRITE]);
 	dup2(pipefd[PIPE_READ], STDIN_FILENO);
 }
-
